Moved copy_recursively function to helpers mod file
